FIxed design responsibility
The current design didn't scale down to iOS (or similar) devices (i.e. didn't make use of the bootstrap-responsive stylesheet). This was due to the missing "initial-scale" meta tag. I added this and furthermore added the utf8-charset tag for HTML5. Besides, I updated the year - there's likely to be a better way of doing it in EJS but this solution works ;)